extracting more ol dependent functions into olViewer
